Improve FMQ pointer corruption diagnostics
Since sometimes an abort stack is the only information
available, it is useful to have suspicious pointer values
in the abort message. For that, it must be generated
by the FMQ implementation itself as the client does not
have access to the pointers.

Because of that, change the parameter introduced by
aosp/3073435 to an optional enum and an error string.

Bug: 338974476
Test: m
Change-Id: Ib9d61c787a9e1cd1197503aed698bc1e7e61bc87