Change FMQ to use WriteTo/TryFromBytes traits, enabling AIDL
FMQ now relies on the traits from go/aidl-with-fmq-rust to
enable usage with AIDL types; add test coverage for these
types now that it's possible.

We can't fuzz the whole space of possible AIDL types due to
monomorphism restrictions on AIDL interfaces, so we test a
handful of AIDL types of differing kinds (union/enum/struct)
defined by hand.

Test: atest fmq_test
Bug: 294698705
Change-Id: Id36604c0ac3b7f2ace85da00b092855663b74686